+++ app/routes/reports.js
@@ -1,9 +1,10 @@
 "use strict";
 var __awaiter = (this && this.__awaiter) || function (thisArg, _arguments, P, generator) {
+    function adopt(value) { return value instanceof P ? value : new P(function (resolve) { resolve(value); }); }
     return new (P || (P = Promise))(function (resolve, reject) {
         function fulfilled(value) { try { step(generator.next(value)); } catch (e) { reject(e); } }
         function rejected(value) { try { step(generator["throw"](value)); } catch (e) { reject(e); } }
-        function step(result) { result.done ? resolve(result.value) : new P(function (resolve) { resolve(result.value); }).then(fulfilled, rejected); }
+        function step(result) { result.done ? resolve(result.value) : adopt(result.value).then(fulfilled, rejected); }
         step((generator = generator.apply(thisArg, _arguments || [])).next());
     });
 };
@@ -11,7 +12,7 @@ Object.defineProperty(exports, "__esModule", { value: true });
 /**
  * レポート出力管理ルーター
  */
-const cinerinoapi = require("@cinerino/api-nodejs-client");
+const cinerinoapi = require("@cinerino/sdk");
 const tttsapi = require("@motionpicture/ttts-api-nodejs-client");
 const express_1 = require("express");
 const reportsController = require("../controllers/reports");
@@ -38,7 +39,7 @@ reportsRouter.get('/sales', (__, res) => {
     });
 });
 // アカウント別レポート出力
-reportsRouter.get('/account', (req, res, next) => __awaiter(this, void 0, void 0, function* () {
+reportsRouter.get('/account', (req, res, next) => __awaiter(void 0, void 0, void 0, function* () {
     try {
         const cognitoCredentials = req.session.cognitoCredentials;
         authClient.setCredentials({