+++ CHANGELOG.md
@@ -13,6 +13,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/).
 - update typescript
 - update tslint
 - @cinerino/api-nodejs-client -> @cinerino/sdk
+- ログイン認証をCognito UIで対応
 
 ### Deprecated