+++ CHANGELOG.md
@@ -18,6 +18,15 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/).
 
 ### Security
 
+## v3.5.0 - 2020-07-22
+
+### Changed
+
+- update typescript
+- update tslint
+- @cinerino/api-nodejs-client -> @cinerino/sdk
+- ログイン認証をCognito UIで対応
+
 ## v3.4.2 - 2020-06-08
 
 ### Changed